Backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
cryptography==46.0.5
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.0
dnspython==2.8.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gspread==6.2.1
gunicorn==25.1.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pymongo==4.16.0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
razorpay==1.4.2
